修改

+++ oap-server/server-starter-boot/server-starter-boot-core/src/main/java/com/piesat/skywalking/vo/FileSystemVo.java
@@ -10,7 +10,7 @@ public class FileSystemVo {
     @ApiModelProperty(value = "磁盘名称")
     private String diskName;
     @ApiModelProperty(value = "磁盘使用率")
-    private double usage;
+    private float usage;
     @ApiModelProperty(value = "磁盘空闲 GB")
-    private double free;
+    private float free;
 }
